+++ tests/TestCase/Utility/BackupManagerTest.php
@@ -48,7 +48,7 @@ class BackupManagerTest extends TestCase
      * Called before every test method
      * @return void
      */
-    protected function setUp(): void
+    public function setUp(): void
     {
         parent::setUp();